pdf-to-video/requirements.txt
Flask==3.1.1
pdfplumber==0.11.7
gTTS==2.5.4
diffusers==0.33.1
transformers==4.52.4
torch==2.5.1
moviepy==1.0.3
Pillow==10.4.0
gunicorn==21.2.0
requests==2.32.4
